Fix rendering of LaTeX formulas with underscores
Due to conflicts with how kramdown parses Markdown first,
underscores are interpreted as _ in LaTeX formulas

The solution is to use $$ delimiters for inline instead of $,
since only $$ are interpreted as math (and thus ignored) by kramdown.
The use of inline vs. display math is decided based on formatting
(e.g. display math if in a new paragraph)